requirements.txt
APScheduler==3.10.1
b64uuid==0.1
backports.zoneinfo==0.2.1
boto3==1.26.92
botocore==1.29.92
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
idna==3.4
jmespath==1.0.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uuid==1.30
wincertstore==0.2
